pyproject.toml
[tool.poetry]
name = "avault"
version = "0.1.1"
description = ""
authors = ["Shoma FUKUDA <fkshom+pypi@gmail.com>"]
homepage = "https://github.com/fkshom"
repository = "https://github.com/fkshom/avault"
documentation = "https://github.com/fkshom/avault"
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    'Programming Language :: Python :: 3',
    "Programming Language :: Python :: 3.7",
    "Topic :: Utilities",
]

[tool.poetry.scripts]
avault = "avault.avault:main"

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.4.1"
ansible = "^2.5"
jinja2 = "^2.11.3"
cryptography = "^3.4.6"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
assertpy = "^1.1"
pytest-watch = "^4.2.0"
pytest-mock = "^3.5.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
